+++ services/httpd/response_logger.go
@@ -5,7 +5,6 @@ import (
 	"net"
 	"net/http"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/influxdata/influxdb/influxql"
@@ -80,7 +79,10 @@ func redactPassword(r *http.Request) {
 // Common Log Format: http://en.wikipedia.org/wiki/Common_Log_Format
 
 // buildLogLine creates a common log format
-// in addition to the common fields, we also append referrer, user agent and request ID
+// in addition to the common fields, we also append referrer, user agent,
+// request ID and response time (microseconds)
+//  ie, in apache mod_log_config terms:
+//     %h %l %u %t \"%r\" %>s %b \"%{Referer}i\" \"%{User-agent}i\"" %L %D
 func buildLogLine(l *responseLogger, r *http.Request, start time.Time) string {
 
 	redactPassword(r)
@@ -99,11 +101,10 @@ func buildLogLine(l *responseLogger, r *http.Request, start time.Time) string {
 
 	userAgent := r.UserAgent()
 
-	fields := []string{
+	return fmt.Sprintf(`%s - %s [%s] "%s %s %s" %s %s "%s" "%s" %s %d`,
 		host,
-		"-",
 		detect(username, "-"),
-		fmt.Sprintf("[%s]", start.Format("02/Jan/2006:15:04:05 -0700")),
+		start.Format("02/Jan/2006:15:04:05 -0700"),
 		r.Method,
 		uri,
 		r.Proto,
@@ -112,10 +113,9 @@ func buildLogLine(l *responseLogger, r *http.Request, start time.Time) string {
 		detect(referer, "-"),
 		detect(userAgent, "-"),
 		r.Header.Get("Request-Id"),
-		fmt.Sprintf("%s", time.Since(start)),
-	}
-
-	return strings.Join(fields, " ")
+		// response time, report in microseconds because this is consistent
+		// with apache's %D parameter in mod_log_config
+		int64(time.Since(start)/time.Microsecond))
 }
 
 // detect detects the first presense of a non blank string and returns it